extract_utils: fix extraction of alternate partitions
The current code checks if partition is in list of normal partitions
before deciding whether to extract it using normal route or firmware
route.

If you want to extract system_ext, but system_ext is not a separate
partition, and it's contained inside system, after trying to extract
system_ext, code will pick system as an alternate partition for
system_ext. Normal partitions will not contain system because we're
either only trying to extract system_ext or because we don't extract
any blob from system so firmware route will be picked, failing to
extract it.

Firmware partitions have no alternate partitions, so flip the condition.

Change-Id: I3b3a65495d8b136a6163d0caee59f9b482bffe2c