extract_utils: Consider path prefixes other than `smali/`
Currently we only decode source code if there are any changed
files in the patch whose path starts with 'smali/', but in the
case of multi dex APKs, apktool will instead give multiple smali
directories in the form 'smali_<name of dex file>'.

Knowing that, a better way to check if the source needs decoding
is to see if any path starts with 'smali', which should deal with
most edge cases.

Signed-off-by: bengris32 <bengris32@protonmail.ch>
Change-Id: I3927fe5bd4aae00dbbd3b1895fb15ca67e33afbe